refactor of classes in api
