+++ src/lib/api/AuthClient/constants.ts
@@ -1,0 +1,7 @@
+export const AUTH_CLIENT_ENDPOINTS = {
+  login: "auth/sign-in",
+  refresh: "auth/refresh",
+  logout: "auth/sign-out",
+  register: "auth/sign-up",
+  session: "auth/session",
+} as const;
