+++ src/lib/api/AuthClient/types.ts
@@ -1,0 +1,4 @@
+export type AuthClientLogoutOptions = {
+  accessToken?: string;
+  refreshToken?: string;
+};
