+++ src/lib/api/SupabaseAuthApiClient/SupabaseAuthApiClient.ts
@@ -5,8 +5,10 @@ import type {
   ForgotPasswordDto,
   ResendConfirmEmailDto,
   ResetPasswordDto,
-} from "../entities";
-import type { IAuthApiClient } from "./IAuthApiClient";
+} from "../../entities";
+import type { IAuthApiClient } from "../IAuthApiClient";
+import { NO_ACTIVE_SESSION_ERROR } from "./constants";
+import { requireRefreshToken, toOtpPayload, trimOrUndefined } from "./utils";
 
 /**
  * Supabase adapter for {@link IAuthApiClient}. Maps DTOs onto
@@ -20,7 +22,7 @@ export class SupabaseAuthApiClient implements IAuthApiClient {
   constructor(private readonly supabase: SupabaseClient) {}
 
   async forgotPassword(data: ForgotPasswordDto): Promise<void> {
-    const redirectTo = data.redirectTo?.trim();
+    const redirectTo = trimOrUndefined(data.redirectTo);
     const { error } = await this.supabase.auth.resetPasswordForEmail(
       data.email,
       redirectTo ? { redirectTo } : undefined,
@@ -29,7 +31,7 @@ export class SupabaseAuthApiClient implements IAuthApiClient {
   }
 
   async resendConfirmEmail(data: ResendConfirmEmailDto): Promise<void> {
-    const redirectTo = data.redirectTo?.trim();
+    const redirectTo = trimOrUndefined(data.redirectTo);
     const { error } = await this.supabase.auth.resend({
       type: "signup",
       email: data.email,
@@ -40,7 +42,7 @@ export class SupabaseAuthApiClient implements IAuthApiClient {
 
   async confirmEmail(data: ConfirmEmailDto): Promise<void> {
     const { error } = await this.supabase.auth.verifyOtp({
-      token_hash: data.tokenHash,
+      ...toOtpPayload(data),
       type: data.type as EmailOtpType,
     });
     if (error) throw error;
@@ -50,19 +52,14 @@ export class SupabaseAuthApiClient implements IAuthApiClient {
   async resetPassword(data: ResetPasswordDto): Promise<void> {
     if ("tokenHash" in data) {
       const { error } = await this.supabase.auth.verifyOtp({
-        token_hash: data.tokenHash,
+        ...toOtpPayload(data),
         type: data.type as EmailOtpType,
       });
       if (error) throw error;
     } else {
-      if (!data.refreshToken) {
-        throw new Error(
-          "SupabaseAuthApiClient.resetPassword: refreshToken is required when using the access-token path.",
-        );
-      }
       const { error } = await this.supabase.auth.setSession({
         access_token: data.accessToken,
-        refresh_token: data.refreshToken,
+        refresh_token: requireRefreshToken(data),
       });
       if (error) throw error;
     }
@@ -70,7 +67,7 @@ export class SupabaseAuthApiClient implements IAuthApiClient {
     const { data: sessionData, error: sessionError } =
       await this.supabase.auth.getSession();
     if (sessionError) throw sessionError;
-    if (!sessionData.session) throw new Error("No active session");
+    if (!sessionData.session) throw new Error(NO_ACTIVE_SESSION_ERROR);
 
     const { error: updateError } = await this.supabase.auth.updateUser({
       password: data.newPassword,
