+++ src/lib/api/SupabaseAuthApiClient/constants.ts
@@ -1,0 +1,3 @@
+export const NO_ACTIVE_SESSION_ERROR = "No active session";
+export const RESET_PASSWORD_REFRESH_TOKEN_REQUIRED_ERROR =
+  "SupabaseAuthApiClient.resetPassword: refreshToken is required when using the access-token path.";
