+++ src/lib/api/SupabaseAuthApiClient/index.ts
@@ -1,0 +1,1 @@
+export * from "./SupabaseAuthApiClient";
