+++ src/lib/api/SupabaseAuthClient/constants.ts
@@ -1,0 +1,4 @@
+export const AUTHENTICATED_SESSION_NOT_FOUND_ERROR =
+  "Authenticated session not found";
+export const NO_ACTIVE_SESSION_ERROR = "No active session";
+export const EMAIL_CONFIRMATION_REQUIRED_ERROR = "Email confirmation required";
