+++ src/lib/api/SupabaseAuthClient/index.ts
@@ -1,0 +1,2 @@
+export * from "./SupabaseAuthClient";
+export * from "./types";
